Reintegrate merge from wsjtx_exp branch.
git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6122 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ signalmeter.h
@@ -16,7 +16,7 @@ public:
   explicit SignalMeter (QWidget * parent = nullptr);
 
 public slots:
-  void setValue (int value);
+  void setValue (float value);
 
 private:
   MeterWidget * m_meter;
